Cassandra/requirements.txt
cassandra-driver
time_uuid
colorama
tabulate